Moved tesseract call to ivre.utils
